+++ src/PredisStore.php
@@ -74,9 +74,7 @@ class PredisStore implements KeyValueStore
         $serialized = null;
 
         try {
-            if ($this->client->exists($key)) {
-                $serialized = $this->client->get($key);
-            }
+            $serialized = $this->client->get($key);
         } catch (Exception $e) {
             throw ReadException::forException($e);
         }
@@ -86,6 +84,39 @@ class PredisStore implements KeyValueStore
         }
 
         return Serializer::unserialize($serialized);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getMultiple(array $keys)
+    {
+        KeyUtil::validateMultiple($keys);
+
+        // Normalize indices of the array
+        $keys = array_values($keys);
+        $values = array();
+        $notFoundKeys = array();
+
+        try {
+            $serializedValues = $this->client->mget($keys);
+        } catch (Exception $e) {
+            throw ReadException::forException($e);
+        }
+
+        foreach ($serializedValues as $i => $serializedValue) {
+            if (null === $serializedValue) {
+                $notFoundKeys[] = $keys[$i];
+            } elseif (!$notFoundKeys) {
+                $values[$keys[$i]] = Serializer::unserialize($serializedValue);
+            }
+        }
+
+        if ($notFoundKeys) {
+            throw NoSuchKeyException::forKeys($notFoundKeys);
+        }
+
+        return $values;
     }
 
     /**